Store XMLTV language codes in EPG DB
- Removed teletext language code from AI block, which was always set to 1
- Added new type EPG_LANG_CODE in new header epg_lang.h;
  Replaced obsolete type XML_LANG_CODE in xmltv/ modules with new type.
- Added language codes to PI_BLOCK for title & description.
- Adapted language detection & selection in UI code to use new type.
- Extended TTX parser to add language attribute to <title>, <sub-title>
  and <desc> tags (only supports EN/DE/FR/IT for now)

+++ epgui/pdc_themes.h
@@ -26,9 +26,9 @@
 // ----------------------------------------------------------------------------
 // Declaration of service interface functions
 //
-void PdcThemeSetLanguage( uchar lang );
+void PdcThemeSetLanguage( uint lang );
 const char * PdcThemeGet( uchar theme );
-const char * PdcThemeGetByLang( uchar theme, uchar lang );
+const char * PdcThemeGetByLang( uchar theme, uint lang );
 const char * PdcThemeGetWithGeneral( uchar theme, const char ** pGeneralStr, bool withUndef );
 bool PdcThemeIsDefined( uchar theme );
 uchar PdcThemeGetCategory( uchar theme );
